requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
transformers==4.41.2
torch==2.3.0
accelerate==0.30.0
pydantic==2.7.1
